requirements.txt
numpy==1.21.2
tqdm==4.62.3
scipy==1.8.0
pandas==1.3.3
torchsummary==1.5.1
idna==3.3 
tabulate==0.8.9 
neptune-client==0.15.2
sty==1.0.0rc2
openpyxl==3.0.9
seaborn==0.11.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
pympler==1.0.1
python-Levenshtein==0.12.2
jellyfish==0.8.2
matplotlib==3.5.1